Allow system to check if user can submit a fetch

+++ backEnd/routes/donationPostCRUD.js
@@ -3,6 +3,12 @@ const router = express.Router();
 const { PrismaClient } = require('../generated/prisma')
 const prisma = new PrismaClient;
 
+const isAuthenticated = (req, res, next) => {
+    if (!req.session.userId) {
+        return res.status(401).json({ error: "You must be logged in to perform this action." });
+    }
+    next();
+};
 
 //get all posts
 router.get('/users/:userId/donations', async (req, res) => {
@@ -30,9 +36,12 @@ router.get('/users/:userId/donations/:postId', async (req, res) => {
 })
 
 //add in a post
-router.post('/users/:userId/donations', async (req, res) => {
+router.post('/users/:userId/donations', isAuthenticated, async (req, res) => {
   const userId = parseInt(req.params.userId);
   const {title, description, photo, useState} = req.body;
+  if(req.session.userId !== userId){
+     return res.status(401).json({ message: "Invalid User" });
+  }
   const newDonationPost = await prisma.donationPost.create({
     data: {
       title,
@@ -46,9 +55,12 @@ router.post('/users/:userId/donations', async (req, res) => {
 })
 
 //deletes post
-router.delete('/users/:userId/donations/:postId', async (req, res) => {
+router.delete('/users/:userId/donations/:postId', isAuthenticated, async (req, res) => {
   const userId = parseInt(req.params.userId);
   const postId = parseInt(req.params.postId);
+  if(req.session.userId !== userId){
+     return res.status(401).json({ message: "Invalid User" });
+  }
   const deletedpost = await prisma.donationPost.delete({
     where: { 
       id: parseInt(postId),
@@ -59,9 +71,12 @@ router.delete('/users/:userId/donations/:postId', async (req, res) => {
 })
 
 //updates post
-router.put('/users/:userId/donations/:postId', async (req, res) => {
+router.put('/users/:userId/donations/:postId', isAuthenticated, async (req, res) => {
   const userId = parseInt(req.params.userId);
   const postId = parseInt(req.params.postId);
+  if(req.session.userId !== userId){
+     return res.status(401).json({ message: "Invalid User" });
+  }
   const {title, description, photo, useState} = req.body;
   const updatedPost = await prisma.donationPost.update({
     where: { id: parseInt(postId),
